Finished 4 basic tests

+++ src/test/java/com/petstoretests/jsondata/models/PetModel.java
@@ -3,7 +3,7 @@ package com.petstoretests.jsondata.models;
 
 import java.util.ArrayList;
 
-public record PetModel(Integer id,
+public record PetModel(Long id,
                        CategoryModel category,
                        String name,
                        ArrayList<String> photoUrls,
